Use async coroutine for dyn attr test
The tests would pass in `GreenMode.Asyncio`
if the read, write and is_allowed methods
are not coroutines.  This does not verify
the problem reported:  specifically, that
dynamic attributes with async coroutines do not
work properly.  With these changes, the tests
now fail as expected for `GreenMode.Asyncio`.

The creation of the dynamic attributes was changed
to use the more convenient `attribute` class instead
of `Attr`.  This also allows all data types to be
more easily tested.  The `initialize_dynamic_attributes`
method is a built-in PyTango feature, called during
device class instantiation, by the `DeviceClass` device
factory.  It is called after `__init__` and
`init_device`, but before the device is exported.

Also moved the test closer to the other dynamic attribute
tests.