setup.py: Define NPY_NO_DEPRECATED_API to avoid warning noise during compilation
During compilation a lot of warnings like

  /home/firma/.pyenv/versions/3.8.6/lib/python3.8/site-packages/numpy/core/include/numpy/npy_1_7_deprecated_api.h:17:2:
  warning: #warning "Using deprecated NumPy API, disable it with " "#define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION" [-Wcpp]
are spit out.

This makes finding real warnings challenging.

Unfortunately the advised method of setting NPY_NO_DEPRECATED_API to
NPY_1_7_API_VERSION does not work as it fails to compile with that. So
pytango is using numpy C interfaces which are gone in numpy 1.7.

Just using "0" suppresses the warnings and is safe according to [1].

[1]: https://github.com/numpy/numpy/issues/11653