Raise when setting non-existent DeviceProxy attr
Setting attributes on the `DeviceProxy` is the
high-level API way of writing to a Tango device
attribute.  If the user of the `DeviceProxy`
misspells the attribute name, they will not know
that the expected attribute hasn't actually been
written to.

Now raise `AttributeError` if trying to set
a non-existent attribute.  This does break
compatibility, but `DeviceProxy` instances
shouldn't be used that way.  The kernel
developers agreed to make this change even though
the next release will only be a minor version bump.
https://gitlab.com/tango-controls/tango-kernel-followup/-/blob/main/2021/2021-07-22/Minutes.md#issues-requiring-discussions

Closes #422