Increase time tolerance in logging tests
There are platform-specific differences in the results.
E.g., on Windows the tests were failing with log message
times in the future by up to 11 milliseconds.  This appears
to be due to differences in the specific functions used
to get the time from the system.  log4tango in cppTango
uses ftime() under Windows.  While CPython uses
GetSystemTimeAsFileTime()

Under Windows a 20 millisecond tolerance was chosen, to
be slightly larger than the 15 ms accuracy reported in this
CPython commit:

https://github.com/python/cpython/commit/09225b73c1460c2b725b009eb8c9cf320fa22d5d