Fix pytest invocation under Windows
The parsing tests were failing to execute under Windows as the command line
in the batch file had unquoted spaces in the test names (nodeid).  Fix
by using quotes, and check pytest exit codes for other problems.

Added `-v` to the pytest command line to make it easier to understand which
tests are being run.