Add more tests for dynamic attributes
This is to cover the cases of bound and unbound
methods, and the variety of ways that dynamic attributes
can be created.  PyTango 9.3.4 had a regression related
to this.

Specifically include a test that matches Sardana's
usage to ensure no regression.

Added a test for static attributes as well, to verify
the default method names work as expected.