Fail if mixed green modes used in device server
The green mode applies to a device server, so all individual
Devices in that device server must specify the same green
mode.  Previously this was only checked when using `MultiDeviceTestContext`,
but now the device server will fail to start whether running
normally, or via a test context.

Devices with unspecified green modes default to the global
green mode.  Default for that is synchronous.

The new tests showed that some combinations of global green
mode and device green mode do not work.  This is not a new
issue.  It is not being resolved now, but could be in future.