Update version for 9.4.0rc1
This is an early release candidate for 9.4.0.  Importantly,
the API change to empty image and spectrum attributes is
still pending.

The `Release.long_version` field has been chosen over the
`Release.version` field in `__init__.py` and `utils.py` since
it include the suffix, like "rc1".  The `long_version` is
already used in `setup.py` as the package version.