Require cppTango>=9.4.1 to import the library
Functionality for empty spectrum attributes is dependent on
cppTango 9.4.1, at least.  Ensure that PyTango is not used
with an earlier version of cppTango.

Fixed a few typos, and other places where we mention cppTango
9.4.0.

Still to do in future MR:  update appveyor and gitlab CI to
use the final 9.4.1 release.