Improve detection of TangoTest server port for testing
After changing the default ORBendPoint host from `""` to `"127.0.0.1`,
the detection of the port to connect the `DeviceProxy` to is unreliable.
This is the GIOP (General Inter-ORB Protocol) port.
Detection is now unreliable because it relied on the GIOP port appearing
as an IPv6 socket.  With the `"127.0.0.1"` host, it is an IPv4 port,  just
like the ZMQ port.  On some platforms, the GIOP port is the first reported
connection, but on others it is second.  Thus test become unreliable.

We now probe all the open ports to determine the protocol:
ZMQ or CORBA GIOP.

The methods are added to test_context.py, as they will be useful for
future work on that improves `DeviceProxy` access when using the
`DeviceTestContext`.  See https://gitlab.com/tango-controls/pytango/-/merge_requests/388.

Adding them to `test_context.py` means that the `psutil` package is now
a installation dependency of PyTango.

Imports are available via `test_utils`, since `DeviceTestContext` is exposed
there as well.