Change event test host to 127.0.0.1, and ephemeral port
The fixed port doesn't seem to be required any more.  Events work without
specifying it.  What is import is a host name.  By default the `DeviceTestContext`
will use an IP adddress.  We change to 127.0.0.1 since we don't need
external access.

Note: there are still some intermittent test failures with this setting,
but that was present before.