Fix device types autocompletion in IDEs
The dynamic inheritance (calling a function) used
for `LatestDeviceImpl` breaks autocompletion in IDEs
like PyCharm.  The `DeviceMeta` class used
to create `Device` also prevents autocompletion.

We fix this via @JeanLucPons's solution:
use static inheritance for `LatestDeviceImpl`,
and add a type hint for `Device`.  We ensure
that `LatestDeviceImpl` is still correct with
a new unit test.

Fixes #512