Change Device metaclass usage to Python 3 syntax
Creating `Device` by instantiating with `DeviceMeta`
and using a type hint is not sufficient for autocompletion
in IDEs like PyCharm.  The methods are shown, but the
declaration and docstring cannot be shown for inherited
methods.

Changing `Device` to use the Python 3 metaclass syntax
fixes the type autocompletion problem and keeps the same
behaviour.  We have a unit test for device inheritance.