Pass device object to all dynamic attr functions
Change the wrapping of the read/write/is_allowed
functions for dynamic attributes so that they
support bound methods and unbound functions
with the same signature.  Now user functions
for dynamic attributes match the way static
attributes work (at the signature level).
The callee will always get a reference to the
device object.

We also add a test case for a non-device class
that hosts the read/write/is_allowed methods.
In this case, the device is also passed to
those calls.

Bonus: removed an unnecessary timeout from one
of the server unit tests.