Fix dynamic attr user methods with same decorators
If we use kwargs to provide user methods for a dynamic
attribute, and two or more of those methods have the same
decorator, it breaks.  The function names created for the
wrapped functions are the same, and the methods are looked
up by name.  They must be unique per device instance.

Made the names unique, and added a test.

Also fixed green mode lookup.  `AttrData` doesn't have those
fields, while `server.attribute` does.  `getattr` on `None`
is safe.