Add test for dynamic attrs toggling green mode accessor
Ensure things work as expected if the green mode executor
wrapping for read/write/is_allowed functions is toggled.

For asyncio, the device won't work properly with green mode
disabled.  Reads raise DevFailed since coroutine is not
awaited.  Writes don't raise DevFailed, but the internal
value doesn't change since coroutine not awaited.

For synchronous and gevent modes, we check the
executor context is as expected.  If using green mode,
then we must be in the executor context.  One more special
case is that the synchronous mode doesn't really have an
executor, so nothing to check.