Update mambaforge images for CI and dev
Use the latest.

In .devcontainer/Dockerfile we install pkg-config from both
apt and conda.  The apt install is in case TangoTest
compilation needs it, while the conda install is for PyTango
compilation.  We could skip the conda installation, but
having it making the container closer to the CI environment.
No need for rsync in the container, so it is removed.